+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - moved setting noTouch to new method from constructor to allow to change behavior on previously created `Draggable`
 - Added `customScroll` property to `Draggable` - allows to specify way to handle customScrolling when using drag on touch devices
 - Added polyfills to work on IE11
+- Added `shift` to dragInfo
 
 # [0.0.2] Fix for chrome - TouchEvent + PointerEvent